Extract utilities for predicates to `QueryPredicates`
